+++ src/Renderer/RenderData.h
@@ -28,9 +28,9 @@ struct PlaneData
         Vertex{glm::vec3{ 0.5f,  0.5f, 0.0f},    glm::vec2{1.0f, 1.0f}},
         Vertex{glm::vec3{-0.5f,  0.5f, 0.0f},    glm::vec2{0.0f, 1.0f}}
     };
-    static constexpr const std::array<std::array<uint16_t, 3>, 2> indices = {
-        std::array<uint16_t, 3>{0, 1, 2}, 
-        std::array<uint16_t, 3>{2, 3, 0}
+    static constexpr const std::array<std::array<uint32_t, 3>, 2> indices = {
+        std::array<uint32_t, 3>{0, 1, 2}, 
+        std::array<uint32_t, 3>{2, 3, 0}
     };
 };
 
@@ -80,45 +80,45 @@ struct FrameData
         Vertex{glm::vec3{ 0.025f,  0.025f, -0.025f},      glm::vec4{0.0f, 0.0f, 1.0f, 1.0f}}    
     };
 
-    static constexpr std::array<std::array<uint16_t, 3>, 12 * 3> indices = {
-            std::array<uint16_t, 3>{8 * 0 + 0, 8 * 0 + 1, 8 * 0 + 2},
-            std::array<uint16_t, 3>{8 * 0 + 2, 8 * 0 + 3, 8 * 0 + 0},
-            std::array<uint16_t, 3>{8 * 0 + 4, 8 * 0 + 0, 8 * 0 + 3},
-            std::array<uint16_t, 3>{8 * 0 + 3, 8 * 0 + 5, 8 * 0 + 4},
-            std::array<uint16_t, 3>{8 * 0 + 6, 8 * 0 + 4, 8 * 0 + 5},
-            std::array<uint16_t, 3>{8 * 0 + 5, 8 * 0 + 7, 8 * 0 + 6},
-            std::array<uint16_t, 3>{8 * 0 + 1, 8 * 0 + 6, 8 * 0 + 7},
-            std::array<uint16_t, 3>{8 * 0 + 7, 8 * 0 + 2, 8 * 0 + 1},
-            std::array<uint16_t, 3>{8 * 0 + 4, 8 * 0 + 6, 8 * 0 + 1},
-            std::array<uint16_t, 3>{8 * 0 + 1, 8 * 0 + 0, 8 * 0 + 4},
-            std::array<uint16_t, 3>{8 * 0 + 5, 8 * 0 + 7, 8 * 0 + 2},
-            std::array<uint16_t, 3>{8 * 0 + 2, 8 * 0 + 3, 8 * 0 + 5},
+    static constexpr std::array<std::array<uint32_t, 3>, 12 * 3> indices = {
+            std::array<uint32_t, 3>{8 * 0 + 0, 8 * 0 + 1, 8 * 0 + 2},
+            std::array<uint32_t, 3>{8 * 0 + 2, 8 * 0 + 3, 8 * 0 + 0},
+            std::array<uint32_t, 3>{8 * 0 + 4, 8 * 0 + 0, 8 * 0 + 3},
+            std::array<uint32_t, 3>{8 * 0 + 3, 8 * 0 + 5, 8 * 0 + 4},
+            std::array<uint32_t, 3>{8 * 0 + 6, 8 * 0 + 4, 8 * 0 + 5},
+            std::array<uint32_t, 3>{8 * 0 + 5, 8 * 0 + 7, 8 * 0 + 6},
+            std::array<uint32_t, 3>{8 * 0 + 1, 8 * 0 + 6, 8 * 0 + 7},
+            std::array<uint32_t, 3>{8 * 0 + 7, 8 * 0 + 2, 8 * 0 + 1},
+            std::array<uint32_t, 3>{8 * 0 + 4, 8 * 0 + 6, 8 * 0 + 1},
+            std::array<uint32_t, 3>{8 * 0 + 1, 8 * 0 + 0, 8 * 0 + 4},
+            std::array<uint32_t, 3>{8 * 0 + 5, 8 * 0 + 7, 8 * 0 + 2},
+            std::array<uint32_t, 3>{8 * 0 + 2, 8 * 0 + 3, 8 * 0 + 5},
 
-            std::array<uint16_t, 3>{8 * 1 + 0, 8 * 1 + 1, 8 * 1 + 2},
-            std::array<uint16_t, 3>{8 * 1 + 2, 8 * 1 + 3, 8 * 1 + 0},
-            std::array<uint16_t, 3>{8 * 1 + 4, 8 * 1 + 0, 8 * 1 + 3},
-            std::array<uint16_t, 3>{8 * 1 + 3, 8 * 1 + 5, 8 * 1 + 4},
-            std::array<uint16_t, 3>{8 * 1 + 6, 8 * 1 + 4, 8 * 1 + 5},
-            std::array<uint16_t, 3>{8 * 1 + 5, 8 * 1 + 7, 8 * 1 + 6},
-            std::array<uint16_t, 3>{8 * 1 + 1, 8 * 1 + 6, 8 * 1 + 7},
-            std::array<uint16_t, 3>{8 * 1 + 7, 8 * 1 + 2, 8 * 1 + 1},
-            std::array<uint16_t, 3>{8 * 1 + 4, 8 * 1 + 6, 8 * 1 + 1},
-            std::array<uint16_t, 3>{8 * 1 + 1, 8 * 1 + 0, 8 * 1 + 4},
-            std::array<uint16_t, 3>{8 * 1 + 5, 8 * 1 + 7, 8 * 1 + 2},
-            std::array<uint16_t, 3>{8 * 1 + 2, 8 * 1 + 3, 8 * 1 + 5},
+            std::array<uint32_t, 3>{8 * 1 + 0, 8 * 1 + 1, 8 * 1 + 2},
+            std::array<uint32_t, 3>{8 * 1 + 2, 8 * 1 + 3, 8 * 1 + 0},
+            std::array<uint32_t, 3>{8 * 1 + 4, 8 * 1 + 0, 8 * 1 + 3},
+            std::array<uint32_t, 3>{8 * 1 + 3, 8 * 1 + 5, 8 * 1 + 4},
+            std::array<uint32_t, 3>{8 * 1 + 6, 8 * 1 + 4, 8 * 1 + 5},
+            std::array<uint32_t, 3>{8 * 1 + 5, 8 * 1 + 7, 8 * 1 + 6},
+            std::array<uint32_t, 3>{8 * 1 + 1, 8 * 1 + 6, 8 * 1 + 7},
+            std::array<uint32_t, 3>{8 * 1 + 7, 8 * 1 + 2, 8 * 1 + 1},
+            std::array<uint32_t, 3>{8 * 1 + 4, 8 * 1 + 6, 8 * 1 + 1},
+            std::array<uint32_t, 3>{8 * 1 + 1, 8 * 1 + 0, 8 * 1 + 4},
+            std::array<uint32_t, 3>{8 * 1 + 5, 8 * 1 + 7, 8 * 1 + 2},
+            std::array<uint32_t, 3>{8 * 1 + 2, 8 * 1 + 3, 8 * 1 + 5},
 
-            std::array<uint16_t, 3>{8 * 2 + 0, 8 * 2 + 1, 8 * 2 + 2},
-            std::array<uint16_t, 3>{8 * 2 + 2, 8 * 2 + 3, 8 * 2 + 0},
-            std::array<uint16_t, 3>{8 * 2 + 4, 8 * 2 + 0, 8 * 2 + 3},
-            std::array<uint16_t, 3>{8 * 2 + 3, 8 * 2 + 5, 8 * 2 + 4},
-            std::array<uint16_t, 3>{8 * 2 + 6, 8 * 2 + 4, 8 * 2 + 5},
-            std::array<uint16_t, 3>{8 * 2 + 5, 8 * 2 + 7, 8 * 2 + 6},
-            std::array<uint16_t, 3>{8 * 2 + 1, 8 * 2 + 6, 8 * 2 + 7},
-            std::array<uint16_t, 3>{8 * 2 + 7, 8 * 2 + 2, 8 * 2 + 1},
-            std::array<uint16_t, 3>{8 * 2 + 4, 8 * 2 + 6, 8 * 2 + 1},
-            std::array<uint16_t, 3>{8 * 2 + 1, 8 * 2 + 0, 8 * 2 + 4},
-            std::array<uint16_t, 3>{8 * 2 + 5, 8 * 2 + 7, 8 * 2 + 2},
-            std::array<uint16_t, 3>{8 * 2 + 2, 8 * 2 + 3, 8 * 2 + 5}
+            std::array<uint32_t, 3>{8 * 2 + 0, 8 * 2 + 1, 8 * 2 + 2},
+            std::array<uint32_t, 3>{8 * 2 + 2, 8 * 2 + 3, 8 * 2 + 0},
+            std::array<uint32_t, 3>{8 * 2 + 4, 8 * 2 + 0, 8 * 2 + 3},
+            std::array<uint32_t, 3>{8 * 2 + 3, 8 * 2 + 5, 8 * 2 + 4},
+            std::array<uint32_t, 3>{8 * 2 + 6, 8 * 2 + 4, 8 * 2 + 5},
+            std::array<uint32_t, 3>{8 * 2 + 5, 8 * 2 + 7, 8 * 2 + 6},
+            std::array<uint32_t, 3>{8 * 2 + 1, 8 * 2 + 6, 8 * 2 + 7},
+            std::array<uint32_t, 3>{8 * 2 + 7, 8 * 2 + 2, 8 * 2 + 1},
+            std::array<uint32_t, 3>{8 * 2 + 4, 8 * 2 + 6, 8 * 2 + 1},
+            std::array<uint32_t, 3>{8 * 2 + 1, 8 * 2 + 0, 8 * 2 + 4},
+            std::array<uint32_t, 3>{8 * 2 + 5, 8 * 2 + 7, 8 * 2 + 2},
+            std::array<uint32_t, 3>{8 * 2 + 2, 8 * 2 + 3, 8 * 2 + 5}
     };
 };
 
@@ -146,7 +146,7 @@ struct SphereData
         constexpr size_t indexCount = (SPHERE_RINGS - 1) * SPHERE_SECTORS * 6;
 
         std::array<Vertex, vertexCount> vertices{};
-        std::array<std::array<uint16_t, 3>, indexCount / 3> indices{};
+        std::array<std::array<uint32_t, 3>, indexCount / 3> indices{};
 
         constexpr float R = 1.0f / static_cast<float>(SPHERE_RINGS - 1);
         constexpr float S = 1.0f / static_cast<float>(SPHERE_SECTORS - 1);
@@ -199,7 +199,7 @@ struct MeshData
     };
 
     std::vector<Vertex> vertices = {};
-    std::vector<std::array<uint16_t, 3>> indices = {};
+    std::vector<std::array<uint32_t, 3>> indices = {};
 };
 
 //------------------------------------------------------
@@ -218,19 +218,19 @@ struct BoxData
     };
 
     std::array<glm::vec3, 8> vertices = {};
-    static constexpr std::array<std::array<uint16_t, 3>, 12> indices = {
-        std::array<uint16_t, 3>{0, 1, 2}, 
-        std::array<uint16_t, 3>{0, 2, 3},
-        std::array<uint16_t, 3>{4, 6, 5}, 
-        std::array<uint16_t, 3>{4, 7, 6},
-        std::array<uint16_t, 3>{0, 3, 7}, 
-        std::array<uint16_t, 3>{0, 7, 4},
-        std::array<uint16_t, 3>{1, 5, 6}, 
-        std::array<uint16_t, 3>{1, 6, 2},
-        std::array<uint16_t, 3>{3, 2, 6}, 
-        std::array<uint16_t, 3>{3, 6, 7},
-        std::array<uint16_t, 3>{0, 4, 5}, 
-        std::array<uint16_t, 3>{0, 5, 1}
+    static constexpr std::array<std::array<uint32_t, 3>, 12> indices = {
+        std::array<uint32_t, 3>{0, 1, 2}, 
+        std::array<uint32_t, 3>{0, 2, 3},
+        std::array<uint32_t, 3>{4, 6, 5}, 
+        std::array<uint32_t, 3>{4, 7, 6},
+        std::array<uint32_t, 3>{0, 3, 7}, 
+        std::array<uint32_t, 3>{0, 7, 4},
+        std::array<uint32_t, 3>{1, 5, 6}, 
+        std::array<uint32_t, 3>{1, 6, 2},
+        std::array<uint32_t, 3>{3, 2, 6}, 
+        std::array<uint32_t, 3>{3, 6, 7},
+        std::array<uint32_t, 3>{0, 4, 5}, 
+        std::array<uint32_t, 3>{0, 5, 1}
     };
 };
